pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cag-bas-rhin"
version = "0.1.0"
description = "Extraction et visualisation de la Carte Archéologique de la Gaule 67/1 (Bas-Rhin)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11",
    "duckdb>=1.1",
    "pyproj>=3.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
ui = [
    "dash>=2.14",
    "dash-bootstrap-components>=1.5",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
cag67 = "src.__main__:cli"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
